api-testing/requirements.txt
fastapi
uvicorn
slowapi
pydantic
requests
upstash-redis
google-generativeai 